+++ pg-libs/pg-passes/pg-optimise/src/lib.rs
@@ -1,0 +1,7 @@
+//! Optimisation passes for Pauli graphs.
+
+mod commuting_sets;
+mod packed_op;
+mod rotation_merging;
+pub use commuting_sets::GroupCommutingOpsPass;
+pub use rotation_merging::RotationMergingPass;