+++ tket-py/tket/metadata.py
@@ -11,10 +11,10 @@ Examples:
     >>> hugr = Hugr()
     >>> node = hugr[hugr.module_root]
     >>>
-    >>> node.metadata[MaxQubitsHint] = 3
+    >>> node.metadata[ExpectedQubitsHint] = 3
     >>> node.metadata[PytketInputParameters] = ["theta", "phi"]
     >>> node.metadata[PytketQubitRegisterNames] = [("q", [0]), ("ancilla", [1])]
-    >>> node.metadata[MaxQubitsHint]
+    >>> node.metadata[ExpectedQubitsHint]
     3
     >>> node.metadata.get(PytketQubitRegisterNames)
     [('q', [0]), ('ancilla', [1])]
@@ -38,10 +38,15 @@ if TYPE_CHECKING:
 
 __all__ = [
     "CircuitRewriteTraces",
+    "ControlledImplementations",
+    "CtrlDaggeredImplementations",
+    "DaggeredImplementation",
+    "ExpectedQubitsHint",
     "HeliosPlatformConfig",
     "HeliosPlatformConfigValue",
     "InlineAnnotation",
     "InlineAnnotationValue",
+    "NumControlQubits",
     "PytketBit",
     "PytketBitRegisterNames",
     "PytketInputParameters",
@@ -62,6 +67,39 @@ PytketQubit: TypeAlias = tuple[str, list[int]]
 #
 # This can be passed to `pytket.unit_id.Bit.from_list`
 PytketBit: TypeAlias = tuple[str, list[int]]
+
+
+class ControlledImplementations(Metadata[list[str]]):
+    """Metadata key for the controlled custom implementations of a function.
+
+    This is used to store the names of the custom controlled implementations of a function, if present.
+    """
+
+    KEY = _metadata.CONTROLLED_IMPLEMENTATIONS
+
+
+class CtrlDaggeredImplementations(Metadata[list[str]]):
+    """Metadata key for the controlled-daggered custom implementations of a function.
+
+    This is used to store the names of the custom controlled-daggered implementations of a function, if present.
+    """
+
+    KEY = _metadata.CTRL_DAGGERED_IMPLEMENTATIONS
+
+
+class NumControlQubits(Metadata[int]):
+    """Metadata key for number of control qubits for the controlled implementations of a function."""
+
+    KEY = _metadata.NUM_CONTROL_QUBITS
+
+
+class DaggeredImplementation(Metadata[str]):
+    """Metadata key for the daggered custom implementations of a function.
+
+    This is used to store the name of the custom daggered implementation of a function, if present.
+    """
+
+    KEY = _metadata.DAGGERED_IMPLEMENTATION
 
 
 class RewriteTraceValue(TypedDict):
@@ -169,7 +207,7 @@ class PytketPhaseExpr(Metadata[str]):
         Use explicit ``tket.global_phase`` operations instead.
     """
 
-    KEY = _metadata.PYTKET_PHASE_EXPR
+    KEY = "TKET1.phase"
 
 
 def _store_pytket_register(value: list[tuple[str, list[int]]]) -> JsonType: